Pass state and function of CollapsibleOutput to its component
